+++ FJ-VDMJ4/src/main/java/com/fujitsu/vdmj/in/statements/INCyclesStatement.java
@@ -72,7 +72,7 @@ public class INCyclesStatement extends INStatement
 				try
 				{
 					ctxt.threadState.setAtomic(true);
-					value = cycles.eval(ctxt).intValue(ctxt);
+					value = cycles.eval(ctxt).intValue(ctxt).longValue();
 				}
 				finally
 				{
