+++ FJ-VDMJ4/src/main/java/com/fujitsu/vdmj/runtime/Breakpoint.java
@@ -140,7 +140,7 @@ public class Breakpoint implements Serializable
 		}
 
 		LexIntegerToken num = (LexIntegerToken)arg;
-		return new INBreakpointExpression(this, cond, num.value);
+		return new INBreakpointExpression(this, cond, num.value.longValue());
 	}
 
 	@Override
